requirements.txt
attr==0.3.2
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==42.0.2
docutils==0.20.1
HTMLParser==0.0.2
importlib_metadata==6.11.0
ipython==8.21.0
ipywidgets==8.1.1
Jinja2==3.1.2
jnius==1.1.0
keyring==24.3.0
lockfile==0.12.2
mock==5.1.0
Pillow==10.1.0
Pillow==10.2.0
protobuf==4.25.2
pyOpenSSL==24.0.0
python_docx==1.1.0
railroad==0.5.0
Sphinx==7.2.6
streamlit==1.29.0
thread==0.1.3
tornado==6.4
truststore==0.8.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
